🚚[Rename] records, record-write, record-detail을 features/record 하위 폴더로 이동
