+++ pages/records/[recordId]/index.tsx
@@ -1,4 +1,4 @@
-import RecordDetail from "@/components/features/record-detail/RecordDetail";
+import RecordDetail from "@/components/features/record/detail/RecordDetail";
 
 export default function RecordDetailPage() {
   return <RecordDetail />;
