+++ pages/records/index.tsx
@@ -1,4 +1,4 @@
-import RecordFeedPage from "@/components/features/records/RecordFeedPage";
+import RecordFeedPage from "@/components/features/record/list/RecordFeedPage";
 import { JSX } from "react";
 
 export default function RecordsPage(): JSX.Element {
