+++ src/components/features/record/detail/RecordDetail.tsx
