+++ src/components/features/record/detail/hooks/mutations/useDeleteRecord.ts
