+++ src/components/features/record/detail/hooks/queries/useFetchRecord.ts
