+++ src/components/features/record/detail/recordDetailContent/RecordDetailContent.stories.tsx
