+++ src/components/features/record/detail/recordDetailContent/RecordDetailContentMain.tsx
