+++ src/components/features/record/detail/recordDetailContent/RecordDetailContentSubInfo.tsx
