+++ src/components/features/record/list/RecordCard/GradientBg.tsx
