+++ src/components/features/record/list/RecordCard/RecordCardContent.tsx
