+++ src/components/features/record/write/lib/mapRecordWriteToCreateBoardInput.ts
